+++ tests/testthat/test_AssayMatrix.R
@@ -8,6 +8,7 @@ test_that("AssayMatrix object can be created from a dgCMatrix", {
   assaymat$from_matrix(mat, index_cols = c("i", "j"), value_col = "counts")
   expect_true(dir.exists(uri))
   expect_s4_class(assaymat$tiledb_array(), "tiledb_array")
+  expect_equal(assaymat$dimnames(), c("i", "j"))
 
   df2 <- assaymat$to_dataframe()
   expect_s3_class(df2, "data.frame")
@@ -31,6 +32,25 @@ test_that("AssayMatrix object can be created from a dgCMatrix", {
     assaymat$get_metadata(SOMA_LEGACY_VALIDITY_KEY),
     SOMA_LEGACY_VALIDITY
   )
+})
+
+test_that("array dimensions can be transposed", {
+  uri <- withr::local_tempdir("assay-matrix-transpose")
+  mat <- create_sparse_matrix_with_string_dims(nrows = 10, ncols = 5)
+
+  assaymat <- AssayMatrix$new(uri = uri, verbose = FALSE)
+  assaymat$from_matrix(mat, index_cols = c("i", "j"), transpose = TRUE)
+  expect_equal(assaymat$dimnames(), c("j", "i"))
+
+  df2 <- assaymat$to_dataframe()
+  # dim values match dim names
+  expect_equal(unique(df2$j), colnames(mat))
+  expect_setequal(unique(df2$i), rownames(mat))
+
+  # recreated matrix is transposed
+  mat2 <- assaymat$to_matrix()
+  expect_equal(rownames(mat2), colnames(mat))
+  expect_setequal(colnames(mat2), rownames(mat))
 })
 
 test_that("Incomplete queries can be completed via batching", {